April 1 data update
There is a new change in data methodology causing numbers to look a bit strange after April 1 update. Starting today, all the county data will be reported as per person's place of residence, not the location where test has been conducted. In addition, data for Zagrebačka county is now reported separately from the data for City of Zagreb. This change is from the official governemtn sources, it is not our methodology change.